% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.r
\name{add_dates}
\alias{add_dates}
\title{Format date information in the data}
\usage{
add_dates(obdata)
}
\arguments{
\item{obdata}{Data frame containing species occurrence records with the following columns: species, Date, gridref, Year, Week, Month, and optionally covnames (see below) and listL}
}
\description{
Format date information in the data
}
