% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rot.r
\name{filter_rot}
\alias{filter_rot}
\title{Filter posterior samples by RoT rules and report removals}
\usage{
filter_rot(posterior_samples)
}
\arguments{
\item{posterior_samples}{A list of length 2 containing posterior samples and sparta run metadata}
}
\value{
A list with elements:
\itemize{
\item posterior_samples: filtered samples
\item metadata: filtered metadata (only kept species)
}
}
\description{
Filter posterior samples by RoT rules and report removals
}
