% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{\code{loadRFile} - function to load Rdata}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{name of the file to load}
}
\value{
file loaded
}
\description{
This function loads sparta Rdata object
}
