% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sparta.r
\name{plot_sparta}
\alias{plot_sparta}
\title{Plot occDet Objects}
\usage{
plot_sparta(occDet, main = occDet$SPP_NAME, reg_agg = NULL)
}
\arguments{
\item{occDet}{An object of class occDet}

\item{main}{The plot title, defaults to the species name}

\item{reg_agg}{The name of a region or region aggregate to plot.
If '' (default) then the overall occupancy estimates are plotted}
}
\description{
Plot occDet Objects
}
