% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samp_occti.r
\name{posterior_samp_occti}
\alias{posterior_samp_occti}
\title{Generate OCCTI posterior samples}
\usage{
posterior_samp_occti(
  occti_outputs,
  save_dir = "occupancy_plots",
  save_plots = TRUE,
  span = 0.75,
  plot_width = 18,
  plot_height = 6,
  n_iter = 1000
)
}
\arguments{
\item{occti_outputs}{A named list of occupancy outputs as returned by \code{load_occti_outputs()}.}

\item{save_dir}{A character string for the directory where plots should be saved (if \code{save_plots = TRUE}).}

\item{save_plots}{Logical. If TRUE, plots will be saved to \code{save_dir}.}

\item{span}{Numeric. The smoothing parameter passed to the \code{loess()} function (e.g. 0.75).}

\item{plot_width}{Numeric. Width of saved plots in inches. Used only if \code{save_plots = TRUE}.}

\item{plot_height}{Numeric. Height of saved plots in inches. Used only if \code{save_plots = TRUE}.}

\item{n_iter}{Integer. Number of iterations to use to sample within mean and standard deviation`.}
}
\value{
A named list where each element is a data frame of LOESS-smoothed occupancy values for one species.
}
\description{
Generates smoothed posterior samples of occupancy from loess model predictions. Optionally saves models to disk.
Also returns a list of LOESS-smoothed values (mean and bounds) per species.
}
