% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_comparison_plots.R
\name{produce_comparison_plots}
\alias{produce_comparison_plots}
\title{Function to produce comparison plots for DEFRA report}
\usage{
produce_comparison_plots(
  occti_loess_samples,
  sparta_occupancy_estimates_folder,
  output_folder = "sparta_occti_comparison",
  sparta_region,
  year_min_for_occti_comp = NULL,
  year_max_for_occti_comp = NULL
)
}
\description{
Function to produce comparison plots for DEFRA report
}
