% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_sparta.R
\name{translate_sparta}
\alias{translate_sparta}
\title{Pivot occupancy posterior samples to long format}
\usage{
translate_sparta(samp_post)
}
\arguments{
\item{samp_post}{A data frame returned by \code{tempSampPost()} or \code{applySamp_single()};
must contain columns named \code{year_XXXX} for each year, as well as \code{species} and \code{iteration}.}
}
\value{
A long-format \code{data.frame} with columns: \code{iteration}, \code{species}, \code{year}, and \code{pred}
}
\description{
Converts a wide-format data frame of occupancy model posterior samples
(e.g., from \code{tempSampPost()}) into a tidy long format, where each row corresponds
to a single prediction for a species-year-iteration combination.
}
\examples{
samp_long <- translate_sparta(result$samp_post)

}
