% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_daisy_chains.r
\name{combine_daisy_chains}
\alias{combine_daisy_chains}
\title{Combine Daisy-Chained Model Outputs Across Iteration Chunks and Chains}
\usage{
combine_daisy_chains(input_path, output_path = "combined_chains")
}
\arguments{
\item{input_path}{Character. Directory with .rdata chain files.}

\item{output_path}{Character. Directory to save combined outputs (default = "combined_chains").}
}
\value{
Invisibly returns vector of saved file paths.
}
\description{
Combines .rdata files from daisy-chained model runs (multiple iteration chunks,
multiple chains per chunk) into a single model object per species.
}
