% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSamps.r
\name{combineSamps}
\alias{combineSamps}
\title{combineSamps}
\usage{
combineSamps(
  species,
  indata,
  keep_iter,
  region,
  sample_n,
  tolerance,
  combined_output,
  minObs,
  scaleObs,
  t0,
  tn,
  filetype,
  iter
)
}
\description{
This function loads and samples posterior distributions from model outputs for a given species.
It is based on the original \code{combineSamps} function used in \code{tempSampPost} from the wrappeR package:
https://github.com/BiologicalRecordsCentre/wrappeR
}
