% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_occti_outputs.r
\name{load_occti_outputs}
\alias{load_occti_outputs}
\title{Load OCCTI Occupancy Outputs}
\usage{
load_occti_outputs(
  results_dir,
  pattern = "*occupancy_output.rds",
  start_year,
  end_year
)
}
\arguments{
\item{results_dir}{A character string specifying the directory containing OCCTI occupancy output \code{.rds} files.}
}
\value{
A named list where each element contains the output of one species and is named by the species.
}
\description{
Loads all OCCTI model output RDS files from a specified directory into a named list.
}
